backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-multipart==0.0.6
